requirements.txt
google-cloud-bigquery==3.* 
requests>=2.28
pandas>=1.5
streamlit>=1.18
python-dotenv
plotly
